curved: support for mappings of arbitrary regularity
This relies on a new implementation of Lagrange interpolation on the
simplex in reference_interpolation.jl, for now implemented only for
triangles but in principle applicable to simplices in any dimension. The
implementation is based on an explicit formula involving barycentric
coordinates, and assuming (as is the case for canonical Pk elements)
equidistant nodes on the simplex.