curved: add torus domain.
The state of curved parametrizations is that using multiple (possibly
nonoverlapping) charts require significant compatibility
at the interface of the charts. This can be difficult to achieve in
practice, so we will not pursue this further. This is why the
`curved_tet_playground_*` scripts are not accurate in that they do not
yield the volume of the region up to quadrature error, while
`curved_tet_torus.jl` in contrast does.

For the above reason this is the last commit which will have non-working
non-overlapping curved tetrahedral meshing code. For release we will
simply assume a single globally smooth parametrization. Future work will
address the more complex multi-chart case.