Curved triangles and tetrahedra via C. Bernardi SINUM 1989. (#123)
* curved: Demo of Bernardi (2d)-curved simplices

* curved: simplify

* Fix bug in curved parametrization

* curved: cleanup/vectorize 2d demo

* curved: initial 3d tet support; only working(?) with three vertices on boundary

* curved: in-progress 3d tet work with single patch

* curved: some small reference interpolation changes

* curved: more in-progress tet work; overlapping patches

* curved: simplify 2D demo, using reference interpolation from Inti

* curved: working 3D demo

Testcase is a spherical geometry with spherical caps removed from poles.
Removal of these poles is so that the curved parametrization is nowhere
degenerate.

* curved: 2d cleanup

* curved: further work on 3D multipatch; +surface area verification via Stokes thm

* curved: add higher-order 3D mappings

* curved: support for mappings of arbitrary regularity

This relies on a new implementation of Lagrange interpolation on the
simplex in reference_interpolation.jl, for now implemented only for
triangles but in principle applicable to simplices in any dimension. The
implementation is based on an explicit formula involving barycentric
coordinates, and assuming (as is the case for canonical Pk elements)
equidistant nodes on the simplex.

* curved: working two parametrizations with cut sphere

* curved: twoparamcutsphere tweaks

* curved: add torus domain.

The state of curved parametrizations is that using multiple (possibly
nonoverlapping) charts require significant compatibility
at the interface of the charts. This can be difficult to achieve in
practice, so we will not pursue this further. This is why the
`curved_tet_playground_*` scripts are not accurate in that they do not
yield the volume of the region up to quadrature error, while
`curved_tet_torus.jl` in contrast does.

For the above reason this is the last commit which will have non-working
non-overlapping curved tetrahedral meshing code. For release we will
simply assume a single globally smooth parametrization. Future work will
address the more complex multi-chart case.

* curved: cleanup (see description)

Revert to before this commit for two-parametrization curved meshes. See
previous commit.

* curved: cleanup

* curved: begin creation of curved `Mesh` objects

* curved: remove hardcoding of on-surface node detection

* curved: tweak

* curved: Build an Inti Mesh object in 2d.

* curved: Build an Inti Mesh object in 3d.

* curved: fix orientation of curved surface elements

* curved: cleanup; avoid modifying original mesh.

* curved: Add curve_mesh() routine.

* curved: allow user selection of smoothness order θ

* format

* curved: support VDIM

* curved: minor fixes. fix curved boundary connectivity

* typo on torus docstring

* better handling of type instaiblity issues in `adaptive_correction`
- trying to create a `zero` element for an accumulator only makes sense if we can infer the type
- otherwise we just evaluate the kernel at some point, and let the Julia type inference system expand the accumulator type if needed

* curved: Various changes.

* curved: improved Lagrange basis computation.

Thank you to Luiz for help with julia generators.

* curved: performance improvements

* curved: cleanup, fix tests

* curved: fix tests?

* curved: 2D multiple volume entities & parametrizations

* curved: cleanup

* curved: efficient mesh setup in 3D

* curved: 3D performance improvements

* curved: 2D performance improvements

* curved: simplify 2D bdry <-> vol entity logic

* start docs on `curve_mesh`

* curved: more docs

* curved: try to make sensible default choice for patch density; fix docs?

* curved: fix docs??

* curved: fix docs??? :(

* update docs on curved mesh

* curved: improve test case stable retrieval of multiple entites, after Luiz

---------

Co-authored-by: maltezfaria <maltezfaria@gmail.com>